moved all xcuitest/general options to general_options.py
